requirements.txt
jinja2
pyyaml
requests~=2.32.3
requests_cache
